dunder-mifflin-play-app/backend/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
Faker
python-dotenv
requests
gunicorn
flask-cors